final notes

+++ app/Http/Controllers/ListingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use id;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -44,7 +45,8 @@ class ListingController extends Controller
         if($request->hasFile('logo')){
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
-    
+        
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -63,6 +65,12 @@ class ListingController extends Controller
 
         //Update listing data
     public function update(Request $request, Listing $listing){
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -88,8 +96,18 @@ class ListingController extends Controller
 
     //Delete Listing   
     public function destroy(Listing $listing) {
+
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Job deleted successfully');
+    }
+
+    //manage listings
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 
 }
